+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/servlets/action/CreateMessageEntryServletTest.java
@@ -2,6 +2,7 @@ package be.orbinson.aem.dictionarytranslator.servlets.action;
 
 import be.orbinson.aem.dictionarytranslator.services.DictionaryService;
 import be.orbinson.aem.dictionarytranslator.services.impl.DictionaryServiceImpl;
+import com.adobe.granite.license.ProductInfoProvider;
 import com.day.cq.replication.Replicator;
 import io.wcm.testing.mock.aem.junit5.AemContext;
 import io.wcm.testing.mock.aem.junit5.AemContextExtension;
@@ -35,6 +36,7 @@ class CreateMessageEntryServletTest {
     @BeforeEach
     void beforeEach() {
         context.registerService(Replicator.class, mock(Replicator.class));
+        context.registerService(ProductInfoProvider.class, mock(ProductInfoProvider.class));
         dictionaryService = context.registerInjectActivateService(new DictionaryServiceImpl());
 
         servlet = context.registerInjectActivateService(new CreateMessageEntryServlet());