+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/servlets/action/DeleteDictionaryServletTest.java
@@ -1,6 +1,7 @@
 package be.orbinson.aem.dictionarytranslator.servlets.action;
 
 import be.orbinson.aem.dictionarytranslator.services.impl.DictionaryServiceImpl;
+import com.adobe.granite.license.ProductInfoProvider;
 import com.day.cq.replication.ReplicationActionType;
 import com.day.cq.replication.ReplicationException;
 import com.day.cq.replication.Replicator;
@@ -21,8 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.*;
 
 @ExtendWith({AemContextExtension.class, MockitoExtension.class})
 class DeleteDictionaryServletTest {
@@ -36,6 +36,7 @@ class DeleteDictionaryServletTest {
     @BeforeEach
     void beforeEach() {
         replicator = context.registerService(Replicator.class, replicator);
+        context.registerService(ProductInfoProvider.class, mock(ProductInfoProvider.class));
         context.registerInjectActivateService(new DictionaryServiceImpl());
         servlet = context.registerInjectActivateService(new DeleteDictionaryServlet());
     }