+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lazy loading of data in UI to prevent performance issues with large dictionaries [#80](https://github.com/orbinson/aem-dictionary-translator/issues/80)
 - Publication of dictionaries recusively publishes all dictionary properties and items [#119](https://github.com/orbinson/aem-dictionary-translator/issues/102)
 - Betty bar navigation does not crash tab [#113](https://github.com/orbinson/aem-dictionary-translator/issues/133)
+- Dictionaries in apps and libs should not be marked as editable in AAMaCS [#132](https://github.com/orbinson/aem-dictionary-translator/issues/132)
 
 ### Changed